btrfs-progs: move send-dump.c to cmds/receive-dump.c
The dump is a command mode of receive.
