+++ Android.mk
@@ -22,7 +22,7 @@ objects := ctree.c disk-io.c kernel-lib/radix-tree.c extent-tree.c print-tree.c 
           common/extent-cache.c extent_io.c volumes.c utils.c repair.c \
           qgroup.c free-space-cache.c kernel-lib/list_sort.c props.c \
           kernel-shared/ulist.c qgroup-verify.c backref.c common/string-table.c task-utils.c \
-          inode.c file.c find-root.c free-space-tree.c help.c send-dump.c \
+          inode.c file.c find-root.c free-space-tree.c help.c cmds/receive-dump.c \
           common/fsfeatures.c kernel-lib/tables.c kernel-lib/raid56.c transaction.c
 cmds_objects := cmds-subvolume.c cmds-filesystem.c cmds-device.c cmds-scrub.c \
                cmds-inspect.c cmds-balance.c cmds-send.c cmds-receive.c \
