+++ cmds/receive-dump.c
@@ -35,7 +35,7 @@
 #include "cmds/commands.h"
 #include "send-utils.h"
 #include "send-stream.h"
-#include "send-dump.h"
+#include "cmds/receive-dump.h"
 
 #define PATH_CAT_OR_RET(function_name, outpath, path1, path2, ret)	\
 ({									\
